+++ .editorconfig
@@ -9,5 +9,6 @@ ij_kotlin_allow_trailing_comma_on_call_site = true
 ktlint_function_naming_ignore_when_annotated_with = Composable
 ktlint_function_signature_rule_force_multiline_when_parameter_count_greater_or_equal_than = 1
 ktlint_standard_annotation = disabled
+ktlint_standard_function-expression-body = disabled
 ktlint_standard_multiline-expression-wrapping = disabled
 ktlint_standard_string-template-indent = disabled
