+++ app/src/main/java/com/adammcneilly/reader/di/RemoteModule.kt
@@ -1,0 +1,62 @@
+package com.adammcneilly.reader.di
+
+import com.adammcneilly.reader.data.openlibrary.OpenLibraryRetrofitAPI
+import com.squareup.moshi.Moshi
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+import javax.inject.Named
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object RemoteModule {
+    private const val OPEN_LIBRARY_BASE_URL = "https://openlibrary.org/"
+
+    @Provides
+    @Singleton
+    fun provideHttpLoggingInterceptor(): HttpLoggingInterceptor {
+        return HttpLoggingInterceptor().apply {
+            level = HttpLoggingInterceptor.Level.BODY
+        }
+    }
+
+    @Provides
+    @Singleton
+    fun provideOkHttpClient(
+        loggingInterceptor: HttpLoggingInterceptor,
+    ): OkHttpClient {
+        return OkHttpClient
+            .Builder()
+            .addInterceptor(loggingInterceptor)
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    @Named("OpenLibraryRetrofit")
+    fun provideOpenLibraryRetrofit(
+        okHttpClient: OkHttpClient,
+    ): Retrofit {
+        return Retrofit
+            .Builder()
+            .baseUrl(OPEN_LIBRARY_BASE_URL)
+            .client(okHttpClient)
+            .addConverterFactory(MoshiConverterFactory.create(Moshi.Builder().build()))
+            .build()
+    }
+
+    @Provides
+    @Singleton
+    fun provideOpenLibraryApiService(
+        @Named("OpenLibraryRetrofit")
+        retrofit: Retrofit,
+    ): OpenLibraryRetrofitAPI {
+        return retrofit.create(OpenLibraryRetrofitAPI::class.java)
+    }
+}
